+++ atjx/src/main/java/com/atjx/model/ReItem.java
@@ -2,6 +2,7 @@ package com.atjx.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -9,7 +10,7 @@ public class ReItem extends BaseObject {
     private int id;
     private String title;
     private String sellPoint;
-    private int price;
+    private BigDecimal price;
     private int num;
     private String barcode;
     private String image;
@@ -17,5 +18,5 @@ public class ReItem extends BaseObject {
     private int status;
     private Date recovered;
     private String recoveredStr;
-    private int del_price;
+    private BigDecimal del_price;
 }
